+++ CHANGELOG.md
@@ -57,7 +57,7 @@ This file is influenced by http://keepachangelog.com/.
 - Support for PaaS instance of govCMS
 - Ability to edit page order
 - Give user option to update page parent
-
+- Ministers index page
 
 ### Changed